fix(knapsack): proper inheritance
