+++ src/tutorials/__init__.py
@@ -1,0 +1,3 @@
+# import pkg_resources
+#
+# __version__ = pkg_resources.require("tutorials")[0].version
