+++ src/tutorials/pattern_generator.py
