+++ src/tutorials/tutorial9_simple_3ratios_gearbox.py
